Also fix user query on the admin UI
Refactored to extract the fix to a utility class also used on the highlight MFA module
