+++ vip-security-boost.php
@@ -18,6 +18,7 @@ declare(strict_types = 1);
 require_once __DIR__ . '/utils/configs.php';
 require_once __DIR__ . '/utils/class-configs.php';
 require_once __DIR__ . '/utils/class-capability-utils.php';
+require_once __DIR__ . '/utils/class-users-query-utils.php';
 require_once __DIR__ . '/email/class-email.php';
 require_once __DIR__ . '/utils/class-constants.php';
 require_once __DIR__ . '/utils/class-logger.php';
